T#775: Extract integrations (Withings + Google OAuth + Gmail) to src/integrations/routes.ts
Phase 1.10 of Backend Modularization (Library #102). Mechanical move, no
logic changes. Largest extraction yet at ~970L.

**Moved to src/integrations/routes.ts (~1024L incl. scaffolding)**:
- OAUTH_KEY constant + encryptToken/decryptToken (AES-256-GCM via crypto module)
- Withings: WITHINGS_CLIENT_ID/SECRET/REDIRECT_URI, withingsSign (HMAC-SHA256),
  getWithingsNonce, ensureFreshWithingsToken, oauthStates Map
- Withings routes: GET /api/oauth/withings/{authorize,callback,status,disconnect},
  GET /api/withings/devices, POST /api/webhooks/withings, POST /api/oauth/withings/sync
- WITHINGS_MEASTYPES Record + syncWithingsMeasurements helper
- Google: GOOGLE_CLIENT_ID/SECRET/REDIRECT_URI/SCOPES, googleOauthStates Map,
  AUDIT_LOG_PATH, AUDIT_LOG_MAX_LINES
- Access control: AUDIT log + allowlist + rate limiter (30 req/min/beast)
- Google PKCE helpers: generateCodeVerifier, generateCodeChallenge
- ensureFreshGoogleToken (refresh helper), checkGoogleAccess
- tagUntrustedContent (Gmail prompt-injection defense), sanitizeMetadata
- Google routes: GET /api/oauth/google/{authorize,callback,status,disconnect},
  GET/POST/DELETE /api/google/access, GET /api/google/audit
- Gmail proxy: getGmailBeast helper, GET /api/google/gmail/{profile,labels,messages,
  messages/:id,threads/:id}
- Withings auto-sync daemon (runWithingsAutoSync) — forward-declared at module level,
  assigned inside registerIntegrationsRoutes, scheduled by initIntegrations()

**Exports**:
- registerIntegrationsRoutes(app, sqlite, { hasSessionAuth, isTrustedRequest, isForgeAuthorized })
- initIntegrations(sqliteDb) — captures sqlite, schedules Withings auto-sync
  (1h interval, 60s warmup) once via integrationsStarted guard

**Imports from**:
- crypto: createHmac, createCipheriv, createDecipheriv, randomBytes (named imports
  to fix 'createHmac not found' under strict mode)
- ../server/security-logger.ts: logSecurityEvent (called from OAuth auth-failure paths)

**Server.ts changes**:
- Added import (after board import)
- Removed Withings auto-sync daemon block (former L5775-5802, ~28L)
- Removed integration block (former L6944-7886 → post-shift L6919-7861, ~943L)
- Replaced with: initIntegrations(sqlite); registerIntegrationsRoutes(app, sqlite, ...);
- Net: server.ts down ~962 lines (was 9,928, now 8,966)

**Pattern**:
- Module-level forward-declared runWithingsAutoSync stub: 'let runWithingsAutoSync = async () => {}'
  initIntegrations() schedules setInterval/setTimeout against this var. registerIntegrationsRoutes()
  re-assigns the var to the actual daemon body (inside the closure with access to sqlite, helpers,
  syncWithingsMeasurements). This preserves init-order semantics: initIntegrations() can be called
  in any order relative to registerIntegrationsRoutes() at server.ts startup.

**Verification**:
- Boot test: server starts clean, '[Withings] Auto-sync enabled (1h interval, first run in 60s)'
  logs from new module
- tsc --noEmit: zero new errors in src/integrations/routes.ts
- No logic changes: 1:1 mechanical move. All auth gates preserved (allowlist check on
  /api/google/access, rate limit on Gmail proxy, Forge-authorized gate on Withings devices)

**Cumulative across T#771-T#775**: server.ts -2997 lines (was 11,963, now 8,966) — 25% reduction
